+++ gui/mainwindow.cpp
@@ -3556,8 +3556,8 @@ MemoryBrowser* MainWindow::openMemoryView(uintptr_t addr) {
     browser->setCodeFinderLauncher([this](uintptr_t addr, bool writesOnly) {
         startCodeFinderForAddress(addr, writesOnly);
     });
-    browser->setAddToList([this](uintptr_t addr) {
-        addressListModel_->addEntry(addr, ValueType::Int32);
+    browser->setAddToList([this](uintptr_t addr, ce::ValueType type) {
+        addressListModel_->addEntry(addr, type);
         statusBar()->showMessage(QString("Added 0x%1 to the cheat table").arg(addr, 0, 16), 3000);
     });